Update e2e helpers file name
