+++ e2e/support/helpers/index.js
@@ -29,5 +29,4 @@ export * from "./e2e-visual-tests-helpers";
 export * from "./e2e-users-helpers";
 export * from "./e2e-viz-settings-helpers";
 export * from "./e2e-models-metadata-helpers";
-export * from "./e2e-scrollbar-helpers";
-export * from "./e2e-ui-elements-styles-assertions-helpers";
+export * from "./e2e-ui-elements-overflow-helpers";
